chore: organize framework package tests
