+++ packages/extension-drag-handle-react/src/DragHandle.registration.test.ts
