+++ packages/react/src/captureDOMSelection.test.ts
